+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Google as assistant provider on dev environment [#4441](https://github.com/rokwire/illinois-app/issues/4441).
 
 ## [6.1.14] - 2024-10-22
 ### Changed